Extract Packing variant reports shared code to Base class
